Update auction manager (#334)
* Concrete AuctionId type.

* Typo fixes.

* Auction manager minor updates.

* Make clippy happy.

* fmt comments.

+++ modules/cdp_treasury/src/mock.rs
@@ -14,7 +14,7 @@ pub type AccountId = u128;
 pub type BlockNumber = u64;
 pub type Amount = i64;
 pub type Share = u64;
-pub type AuctionId = u64;
+pub type AuctionId = u32;
 
 pub const ALICE: AccountId = 0;
 pub const BOB: AccountId = 1;
