+++ modules/honzon/src/mock.rs
@@ -43,7 +43,7 @@ impl_outer_origin! {
 
 pub type AccountId = u128;
 pub type BlockNumber = u64;
-pub type AuctionId = u64;
+pub type AuctionId = u32;
 
 pub const ALICE: AccountId = 1;
 pub const BOB: AccountId = 2;
